image_classification_animals/requirements.txt
tensorflow>=2.14
keras>=2.14
numpy
pandas
matplotlib
scikit-learn
opencv-python
Pillow